update and fix all

+++ app/Http/Controllers/CompanyController.php
@@ -10,7 +10,9 @@ use Yajra\DataTables\Facades\DataTables;
 use Illuminate\Contracts\Encryption\DecryptException;
 use Illuminate\Support\Facades\Crypt;
 use DB;
+use Auth;
 use Alert;
+use App\Models\Agreement;
 
 class CompanyController extends Controller
 {
@@ -127,7 +129,16 @@ class CompanyController extends Controller
         DB::beginTransaction();
 
         try {
-            Company::find($id)->delete();
+            $checkAgree = Agreement::where('company_id',$id)->first();
+
+            if($checkAgree) {
+                DB::rollback();
+                Alert::info('CANT','Deletion not allowed, has been used');
+
+                return back();
+            }else{
+                Company::find($id)->delete();
+            }
             
             DB::commit();
             Alert::success('SUCCEED','Data deletion has been successful');
@@ -151,31 +162,33 @@ class CompanyController extends Controller
             ->addColumn('division','{{$model->divisions->name}}')
             ->addColumn('action',function($model){
             $btn = '<div class="d-flex justify-content-center"><a href="/company/show/'.Crypt::encryptString($model->id).'" class="btn btn-primary btn-sm m-1"><i class="ti ti-eye"></i></a><a href="/company/edit/'.Crypt::encryptString($model->id).'" class="btn btn-success btn-sm m-1" style="color: white"><i class="ti ti-pencil"></i></a><button type="button" class="btn btn-danger btn-sm m-1" data-bs-toggle="modal" data-bs-target="#modalDelete'.$model->id.'"><i class="ti ti-trash"></i></button></div>
-                    <!-- Modal -->
-                        <div class="modal fade" id="modalDelete'.$model->id.'" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-                        <div class="modal-dialog">
-                            <div class="modal-content">
-                            <div class="modal-header">
-                            </div>
-                            <div class="modal-body">
-                                <div class="row">
-                                    <div class="col-4">
-                                        <img src="https://cdn3d.iconscout.com/3d/premium/thumb/alert-notification-8400231-6672344.png" style="width: 100%;"/>
+                <!-- Modal -->
+                    <div class="modal fade" id="modalDelete'.$model->id.'" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
+                    <div class="modal-dialog">
+                        <div class="modal-content">
+                        <div class="modal-header">
+                        </div>
+                        <div class="modal-body">
+                            <div class="row">
+                                <div class="col-4">
+                                    <img src="https://cdn3d.iconscout.com/3d/premium/thumb/alert-notification-8400231-6672344.png" style="width: 100%;"/>
+                                </div>
+                                <div class="col-8 text-sm">
+                                    <div class="from group mb-3">
+                                    <h5 class="mt-2">'.$model->name.'</h5>
                                     </div>
-                                    <div class="col-8 text-center align-self-center text-uppercase">
-                                        <span>Konfirmasi Penghapusan Data Dari Layanan Legal</span><br>
-                                        <span class="fw-bold">'.$model->name.'</span>
-                                    </div>
+                                    <span style="font-size: 10px;" class="text-lowercase">Proses penghapusan data akan dilakukan secara permanen dari layanan legal, yakin '.Auth::user()->name.' ingin menghapus ?</span>
                                 </div>
                             </div>
-                            <div class="modal-footer">
-                                <button type="button" class="btn btn-secondary text-sm" data-bs-dismiss="modal">Batal</button>
-                                <a href="/company/delete/'.$model->id.'" class="btn btn-danger text-sm">Hapus</a>
-                            </div>
-                            </div>
+                        </div>
+                        <div class="modal-footer">
+                            <button type="button" class="btn btn-secondary text-sm" data-bs-dismiss="modal" style="border-radius: 50%;"><i class="bi bi-x fa-lg"></i></button>
+                            <a href="/company/delete/'.$model->id.'" class="btn btn-primary text-sm text-capitalize" style="border-radius: 50%;"><i class="bi bi-check-lg fa-lg"></i></a>
                         </div>
                         </div>
-                        ';
+                    </div>
+                    </div>
+                    ';
             return $btn;
         })
         ->rawColumns(['action'])
